% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM.R
\name{ADMM}
\alias{ADMM}
\title{ADMM}
\usage{
ADMM(
  x,
  y,
  family = c("gaussian", "binomial", "multinomial", "poisson"),
  intercept = TRUE,
  center = !inherits(x, "sparseMatrix"),
  scale = c("l2", "l1", "sd", "none"),
  alpha = c("path", "estimate"),
  lambda = c("bh", "gaussian", "oscar"),
  alpha_min_ratio = if (NROW(x) < NCOL(x)) 0.01 else 1e-04,
  path_length = if (alpha[1] == "estimate") 1 else 20,
  q = 0.1 * min(1, NROW(x)/NCOL(x)),
  screen = TRUE,
  screen_alg = c("strong", "previous"),
  tol_dev_change = 1e-05,
  tol_dev_ratio = 0.995,
  max_variables = NROW(x),
  opt_algo = c("lbfgs", "bfgs", "nr"),
  max_passes = 1e+06,
  tol_abs = 1e-05,
  tol_rel = 1e-04,
  tol_rel_gap = 1e-05,
  tol_infeas = 0.001,
  diagnostics = FALSE,
  verbosity = 0,
  sigma,
  n_sigma,
  lambda_min_ratio
)
}
\arguments{
\item{x}{the design matrix, which can be either a dense
matrix of the standard \emph{matrix} class, or a sparse matrix
inheriting from \link[Matrix:sparseMatrix]{Matrix::sparseMatrix}. Data frames will
be converted to matrices internally.}

\item{y}{the response, which for \code{family = "gaussian"} must be numeric; for
\code{family = "binomial"} or \code{family = "multinomial"}, it can be a factor.}

\item{family}{model family}

\item{intercept}{whether to fit an intercept}

\item{alpha}{scale for regularization path: either a decreasing numeric
vector (possibly of length 1) or a character vector; in the latter case,
the choices are:
\itemize{
\item \code{"path"}, which computes a regularization sequence
where the first value corresponds to the intercept-only (null) model and
the last to the almost-saturated model, and
\item \code{"estimate"}, which estimates a \emph{single} \code{alpha}
using Algorithm 5 in Bogdan et al. (2015).
}}

\item{lambda}{either a character vector indicating the method used
to construct the lambda path or a numeric non-decreasing
vector with length equal to the number
of coefficients in the model; see section \strong{Regularization sequences}
for details.}

\item{opt_algo}{Algorithm to use for optimizing second order approximation for beta update step.}

\item{max_passes}{maximum number of passes (outer iterations) for solver}

\item{tol_abs}{absolute tolerance criterion for ADMM solver}

\item{tol_rel}{relative tolerance criterion for ADMM solver}

\item{diagnostics}{whether to save diagnostics from the solver}

\item{verbosity}{level of verbosity for displaying output from the
program. Not completely developed. Use 3 just for now.}
}
\description{
ADMM
}
