% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary_covtype}
\alias{binary_covtype}
\title{Binary Covtype}
\format{
An object of class \code{list} of length 2.
}
\usage{
binary_covtype
}
\description{
Binary Covtype
}
\section{Preprocessing}{

Data has been preprocesses so as to have a binary response variable
}

\keyword{datasets}
