% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBenchmarks}
\alias{getBenchmarks}
\title{Runs all the solvers on (x, y) training data with SLOPE parameter alpha,
prints the total time in each case, and returns the merged data frame.}
\usage{
getBenchmarks(
  x,
  y,
  family = c("gaussian", "binomial", "multinomial", "poisson"),
  alpha = 0.01,
  path_length = 1
)
}
\arguments{
\item{x}{the design matrix, which can be either a dense
matrix of the standard \emph{matrix} class, or a sparse matrix
inheriting from \link[Matrix:sparseMatrix]{Matrix::sparseMatrix}. Data frames will
be converted to matrices internally.}

\item{y}{the response, which for \code{family = "gaussian"} must be numeric; for
\code{family = "binomial"} or \code{family = "multinomial"}, it can be a factor.}

\item{family}{model family}

\item{alpha}{parameter used for SLOPE regularization}

\item{path_length}{The regularization path length. By default, it is one.
'alpha' is ignored if path_length is not 1, in which case, list of alpha
and the length of the list is returned.}
}
\description{
Runs all the solvers on (x, y) training data with SLOPE parameter alpha,
prints the total time in each case, and returns the merged data frame.
}
