% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeFits}
\alias{mergeFits}
\title{FISTA}
\usage{
mergeFits(fit1, fit2)
}
\arguments{
\item{fit1}{Output of any of the algorithms (FISTA, ADMM etc)
that you want to compare}

\item{fit2}{Output of any of the algorithms (FISTA, ADMM etc)
that you want to compare with.}
}
\description{
FISTA
}
