% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM.R
\name{ADMM}
\alias{ADMM}
\title{ADMM}
\usage{
ADMM(
  x,
  y,
  family = c("gaussian", "binomial", "multinomial", "poisson"),
  intercept = TRUE,
  center = !inherits(x, "sparseMatrix"),
  scale = c("l2", "l1", "sd", "none"),
  alpha = c("path", "estimate"),
  lambda = c("bh", "gaussian", "oscar"),
  alpha_min_ratio = if (NROW(x) < NCOL(x)) 0.01 else 1e-04,
  path_length = if (alpha[1] == "estimate") 1 else 20,
  q = 0.1 * min(1, NROW(x)/NCOL(x)),
  screen = TRUE,
  screen_alg = c("strong", "previous"),
  tol_dev_change = 1e-05,
  tol_dev_ratio = 0.995,
  max_variables = NROW(x),
  opt_algo = c("lbfgs", "bfgs", "nr"),
  max_passes = 1e+06,
  tol_abs = 1e-05,
  tol_rel = 1e-04,
  diagnostics = FALSE,
  verbosity = 0
)
}
\arguments{
\item{x}{the design matrix, which can be either a dense
matrix of the standard \emph{matrix} class, or a sparse matrix
inheriting from \link[Matrix:sparseMatrix]{Matrix::sparseMatrix}. Data frames will
be converted to matrices internally.}

\item{y}{the response, which for \code{family = "gaussian"} must be numeric; for
\code{family = "binomial"} or \code{family = "multinomial"}, it can be a factor.}

\item{family}{model family}

\item{intercept}{whether to fit an intercept}

\item{center}{whether to center predictors or not by their mean. Defaults
to \code{TRUE} if \code{x} is dense and \code{FALSE} otherwise.}

\item{scale}{type of scaling to apply to predictors.
\itemize{
\item \code{"l1"} scales predictors to have L1 norms of one.
\item \code{"l2"} scales predictors to have L2 norms of one.#'
\item \code{"sd"} scales predictors to have a population standard deviation one.
\item \code{"none"} applies no scaling.
}}

\item{alpha}{scale for regularization path: either a decreasing numeric
vector (possibly of length 1) or a character vector; in the latter case,
the choices are:
\itemize{
\item \code{"path"}, which computes a regularization sequence
where the first value corresponds to the intercept-only (null) model and
the last to the almost-saturated model, and
\item \code{"estimate"}, which estimates a \emph{single} \code{alpha}
using Algorithm 5 in Bogdan et al. (2015).
}}

\item{lambda}{either a character vector indicating the method used
to construct the lambda path or a numeric non-decreasing
vector with length equal to the number
of coefficients in the model; see section \strong{Regularization sequences}
for details.}

\item{alpha_min_ratio}{smallest value for \code{lambda} as a fraction of
\code{lambda_max}; used in the selection of \code{alpha} when \code{alpha = "path"}}

\item{path_length}{length of regularization path; note that the path
returned may still be shorter due to the early termination criteria
given by \code{tol_dev_change}, \code{tol_dev_ratio}, and \code{max_variables}.}

\item{q}{parameter controlling the shape of the lambda sequence, with
usage varying depending on the type of path used and has no effect
is a custom \code{lambda} sequence is used.}

\item{screen}{(currently inactive) whether to use predictor screening rules
(rules that allow some predictors to be discarded prior to fitting),
which improve speed greatly when the number of predictors is larger than
the number of observations.}

\item{screen_alg}{(currently inactive) what type of screening algorithm to use.
\itemize{
\item \code{"strong"} uses the set from the strong screening rule and check
against the full set
\item \code{"previous"} first fits with the previous active set, then checks
against the strong set, and finally against the full set if there are
no violations in the strong set
}}

\item{tol_dev_change}{the regularization path is stopped if the
fractional change in deviance falls below this value; note that this is
automatically set to 0 if a alpha is manually entered}

\item{tol_dev_ratio}{the regularization path is stopped if the
deviance ratio
\eqn{1 - \mathrm{deviance}/\mathrm{(null-deviance)}}{1 - deviance/(null deviance)}
is above this threshold}

\item{max_variables}{criterion for stopping the path in terms of the
maximum number of unique, nonzero coefficients in absolute value in model.
For the multinomial family, this value will be multiplied internally with
the number of levels of the response minus one.}

\item{opt_algo}{Algorithm to use for optimizing second order approximation for beta update step.}

\item{max_passes}{maximum number of passes (outer iterations) for solver}

\item{tol_abs}{absolute tolerance criterion for ADMM solver}

\item{tol_rel}{relative tolerance criterion for ADMM solver}

\item{diagnostics}{whether to save diagnostics from the solver}

\item{verbosity}{level of verbosity for displaying output from the
program. Not completely developed. Use 3 just for now.}
}
\description{
ADMM
}
