% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeFits}
\alias{mergeFits}
\title{Merges all the fits and returns a dataframe containing losses and time}
\usage{
mergeFits(fits, cutoff_time = 10000)
}
\arguments{
\item{fits}{List of outputs of any of the algorithms (FISTA, ADMM etc)}
}
\description{
Merges all the fits and returns a dataframe containing losses and time
}
